backend/requirements.txt
celery==5.4.0
django==5.1.3
django-redis==5.4.0
djangorestframework==3.15.2
drf-spectacular==0.27.2
gunicorn==23.0.0
openpyxl==3.1.5
pandas==2.2.3
psycopg2-binary==2.9.10
pytest-django==4.9.0
python-dotenv==1.0.1
redis==5.2.0
